pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lazygit-ai"
version = "0.1.0"
description = "AI-powered commit message generator that integrates seamlessly with LazyGit"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "lazygit-ai contributors", email = "hello@lazygit-ai.dev"}
]
keywords = ["git", "lazygit", "ai", "commit", "cli", "developer-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Version Control :: Git",
]
requires-python = ">=3.8"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "textual>=0.40.0",
    "gitpython>=3.1.0",
    "pyperclip>=1.8.0",
    "toml>=0.10.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "requests>=2.31.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
lazygit-ai = "lazygit_ai.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/lazygit-ai"
Repository = "https://github.com/yourusername/lazygit-ai"
Documentation = "https://github.com/yourusername/lazygit-ai#readme"
Issues = "https://github.com/yourusername/lazygit-ai/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["lazygit_ai*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=lazygit_ai --cov-report=term-missing --cov-report=html" 